- layout fix - news fetch change to only when haven't been loaded
