+++ src/app/app-providers.tsx
@@ -13,6 +13,8 @@ html, body {
 
 #root {
   height: 100%;
+  display: flex;
+  flex-direction: column;
 }
 `;
