+++ src/app/news/guards/news-guard.tsx
@@ -1,25 +1,19 @@
-import { PropsWithChildren, useEffect } from 'react';
+import { PropsWithChildren } from 'react';
 import { useSelector } from 'react-redux';
-import { useDispatch } from 'src/app/core/redux/redux-hooks';
 import { Results } from 'src/app/shared/components/results';
+import { useFetchNews } from '../hooks/use-fetch-news';
 import {
-  getNewsThunk,
   selectNewsError,
   selectNewsLoaded,
   selectNewsLoading,
 } from '../state/news.state';
 
 export function NewsGuard({ children }: PropsWithChildren) {
-  const dispatch = useDispatch();
   const loaded = useSelector(selectNewsLoaded);
   const loading = useSelector(selectNewsLoading);
   const error = useSelector(selectNewsError);
 
-  useEffect(() => {
-    if (!loading && !loaded) {
-      dispatch(getNewsThunk());
-    }
-  }, [loaded, loading, dispatch]);
+  useFetchNews();
 
   return (
     <Results loading={loading || !loaded} error={error} data={loaded}>
